microservices/authorization/pyproject.toml
[tool.poetry]
name = "authorization"
version = "0.1.0"
description = ""
authors = ["rtrive <riccardo.trivellato@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "*"
pydantic = "*"
python3-indy = "*"
SQLAlchemy = "*"
psycopg2-binary = "*"
pycryptodome = "*"
SQLAlchemy-Paginator = "*"
aws-lambda-powertools = "^1.14.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
flake8 = "^3.9.1"
black = "^20.8b1"
coverage = "^5.5"
mypy = "^0.812"
moto = {extras = ["sqs", "secretsmanager"], version = "^2.0.5"}
Faker = "^8.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
